+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog follows the patterns described here: https://keepachangelog.com/e
 ## 0.2.0
 ### added
 - Added `head_classes`, `body_classes` & `foot_classes` optional properties to the `Hero` component. These extra classes will be added to the stanard classes assigned to each of the corresponding hero component sections.
+- Added `navburger_classes` prop to the `Navbar` component. These extra classes will be added to the standard classes for the `navbar-burger` element.
 
 ## 0.1.6
 ### changed